Fixed a problem with setting the axis type for IrregularPlot objects. (showed up as a bug in example pr04.)
Added calls to free workspaces in Destroy methods for VectorPlot,
StreamlinePlot, and the MapPlot data handler objects.

Made some changes to improve performance in the Map datahandlers: more
attempt is made to preserve the contents of the areamaps between calls, so
that, for instance, changing a color or fill type resource does not
always cause the areamap to be recreated from scratch.

Also modified the canonical ordering of the areas for the v41 database.
Now, the names as returned in the mpAreaNames resource are ordered
alphabetically based on their full path in the hierarchy. This arrangement
is similar to but not exactly the same as an alphbetical listing of
the full names returned by the LLU mpfnme routine. Also modified the
returned names to include enough of the parent hierarchy to ensure that
each name is unique. This will be further explained in the documentation.

This is intended to be the last checkin of new MapPlot stuff for 4.1,
except for bug fixes.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5034 10fe2890-8a40-4696-8b1d-7ccbb19c398e