Added new MapPlot resources to support access to the new Ezmap grid parameters: 	mpGridLatSpacingF : spacing for grid latitude lines in degrees 	mpGridLonSpacingF : spacing for grid longitude lines in degrees 	mpGridMaxLatF     : the maximum latitude (N or S) that a latitude                             line is to appear 	mpGridPolarLonSpacingF : the spacing for longitude lines that                             continue to the poles beyond the maximum 			    latitude lines (only applies to projections that 			    map the poles to single points).
The old resource, mpGridSpacingF, still exists, and, when set, overrides
any setting of mpGridLatSpacingF and mpGridLonSpacingF. In this case,
mpGridLatSpacingF and mpGridLonSpacingF will both be set to the value of
mpGridSpacingF.

Also added code that causes mpAreaMaskingOn to be set to True automatically
if mpMaskAreaSpecifiers is set without setting mpAreaMaskingOn. This is
intended as a simple convenience: hopefully it will not break any existing
code.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5036 10fe2890-8a40-4696-8b1d-7ccbb19c398e