Changed the following lines in MapPlotInitialize from:
	if (! Mpp->area_masking_on_set && Mpp->mask_area_specs )
                Mpp->area_masking_on = True;
	else
                Mpp->area_masking_on = False;

To:

if ((! Mpp->area_masking_on_set )&& (Mpp->mask_area_specs != NULL))
                Mpp->area_masking_on = True;

The problem occurs when area_mask_on_set is set and the area_map_specs  is
also set. This caused MapPlotInitialize to turn off the masking which was
not correct. Since the the mv41BuildFillDrawList handles the situation where
Mpp->mask_area_specs  is not set then the only condition that needs to be
tested for in mapPlotInitialize is when the Mpp->mask_area_specs field is
set and area_masking_on is not set.

	-ethan


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5040 10fe2890-8a40-4696-8b1d-7ccbb19c398e