addfile: minor changes
app: added callback to detect when the _NgSelectedWork was being destroyed
so that it could be reassigned to another workstation.

browse: changed the name of the delete button to Hide; hopefully this will
 be less confusing. Actually I intend to move all button names to the
resource file, but haven't had the chance yet.

go: added a print menu item and code to bring up the Print dialog.

restree: a hopefully temporary hack to eliminate certain resources
being visible when they shouldn't be.

vcrcontrol: added a name parameter to the creation routine.

xapp: stuff to implement the printing facility

print: this is the new print dialog. It supports direct output to the
printer, as well as to several file formats, including the various PS
flavors, NCGM, and GIF.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5111 10fe2890-8a40-4696-8b1d-7ccbb19c398e