Adding rgb database - NAMED COLORS!
The following copyright needs to be put in the documentation somewhere:

****************************************************************************
        Copyright    Massachusetts Institute of Technology    1985
Permission to use, copy, modify, distribute, and sell this software and its
documentation for any purpose is hereby granted without fee, provided that
the above copyright notice appear in all copies and that both that
copyright notice and this permission notice appear in supporting
documentation, and that the name of M.I.T. not be used in advertising or
publicity pertaining to distribution of the software without specific,
written prior permission.  M.I.T. makes no representations about the
suitability of this software for any purpose.  It is provided "as is"
without express or implied warranty.
****************************************************************************


To see the named colors that are valid see: $(NCARG_DATABASE)/rgb.txt

To add named colors to that database:
	ncarg/ncarg2d/src/db/rgb.txt - then "make;make install"

This change makes named colors available for all resources of type:
	NhlTColorIndex, NhlTColorIndexFullEnum,
	NhlTColorIndexGenArray, NhlTColorIndexFullEnumGenArray,
	NhlTColorMap, NhlTColorDefinition

To set a resource of type NhlTColorIndex, NhlTColorIndexFullEnum
	valid "from" values are:

	*	an array of 3 flt/dbl's that will be interpreted as
		rgb intensities - the closest color currently defined
		will be selected.

	*	A string in the following formats:
		(/n,n,n/)
		colorname	- (from rgb.txt)
		#rgb
		#rrggbb
		#rrrgggbbb
		enumname	- (transparent,background...)

To set a resource of type NhlTColorIndexGenArray,
	NhlTColorIndexFullEnumGenArray valid "from" values are:

	*	a 2-d array[*][3] of flt/dbl's that will be interpreted as
		rgb intensities - the closest color currently defined
		will be selected.

	*	An array of strings with each element in the following formats:
		(/n,n,n/)	- (float intensities)
		colorname	- (from rgb.txt)
		#rgb		- ("X" color format - hex digits)
		#rrggbb
		#rrrgggbbb
		enumname	- (transparent,background...)

	*	A single string in resource file pseudo-array format:
		(/ (/n,n,n/),
		   colorname,
		   #rgb,
		   #rrggbb,
		   #rrrgggbbb,
		   enumname,
		   ...		/)

To set a resource of type NhlTColorMap valid "from" values are:

	*	a 2-d array[*][3] of flt/dbl's that will be interpreted as
		rgb intensities - the closest color currently defined
		will be selected.

	*	An array of strings with each element in the following formats:
		(/n,n,n/)	- (float intensities)
		colorname	- (from rgb.txt)
		#rgb		- ("X" color format - hex digits)
		#rrggbb
		#rrrgggbbb

	*	A single string in resource file pseudo-array format:
		(/ (/n,n,n/),
		   colorname,
		   #rgb,
		   #rrggbb,
		   #rrrgggbbb,
		   ...		/)

	*	A single string with a colormap name:
		(default,cyclic,gscyclic...)

To set a resource of type NhlTColorDefinition
	(wkBackgroundColor,wkForegroundColor) valid "from" values are:

	*	an array of 3 flt/dbl's that will be interpreted as
		rgb intensities.

	*	A string in the following formats:
		(/n,n,n/)	- (float intensities)
		colorname	- (from rgb.txt)
		#rgb		- ("X" color format - hex digits)
		#rrggbb
		#rrrgggbbb

jeff


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5163 10fe2890-8a40-4696-8b1d-7ccbb19c398e