Disabled saving MapPlot areamaps to temporary files, because it was causing problems. Modified segment code to handle running out of available segments more gracefully (this required some modification of all view classes as well as Segment.c).
Base.c: Modified the default ordering when children are added
to any Base class object. New children are now put at the end of the
list. For view children of a Workstation this means that the last added
child is the last drawn and therefore appears on top. Still to come is
code that allows the order to be manipulated.

PlotManager.c: added code to do the OverlayStatus callback for all overlays
when a base plot is destroyed.

TextItem.c: added code to distinguish a viewport 'move' from a 'resize'.
turns out that it wasn't really necessary, but it still might add a slight
performance enhancement and I don't think it hurts anything.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5185 10fe2890-8a40-4696-8b1d-7ccbb19c398e