ContourPlot: fixed a redraw-problem for ContourPlots overlaid on MapPlots
MapV40DataHandler: added check to ensure that areamaps are regenerated
when they need to be.

MapV41DataHandler: added areamap check, and fixed a problem with the
strip area group that could result in map fill disappearing when a filled
ContourPlot is overlaid.

PlotManager: added code to the PlotManagerGetBB method to handle special
situations involving zone 0 and 1 annotations.

VectorPlot: fixed an error in the data extent specification when overlaid
on a MapPlot.

Workspace: added new interface routine that allows an object to check whether
the workspace data is still intact.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5187 10fe2890-8a40-4696-8b1d-7ccbb19c398e