This check-in extends the implementation of the viewable object tree in the main window to include representation of Annotations and Overlays. Unlimited depth for annotation plots is supported. Deletions and workstation changes are accounted for. It is still crude. No support for changing the order of plots yet.
It also implements interactive manipulation of View objects in an
XWorkstation window. Plots can be moved and resized using the mouse.
Plots that become exposed because another plot is moved or resized are
redrawn. Automatic updating of the workstation window also take place when
an overlay, annotation, create, delete, or workstation command is issued
from the ncl editor or from a script. This is an option that for the most
part can be turned off, though for now it is on by default. Note that
setvalues commands issued from the ncl editor or a script do not cause
updating. This is because there are still unresolved issues that relate to
difficulties in separating SetValues calls issued internally by the HLU
library from user-initiated SetValues calls. Added "Draw All" and "Clear All"
menu items to the X workstation view menu. Also added an "Draw single view
only" option to the option menu.

When an XWorkstation is created via the Create submenu under File, the
preview workstation no longer pops up.

There are a number of miscellaneous bug fixes included as well.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5202 10fe2890-8a40-4696-8b1d-7ccbb19c398e