Had some funky stuff going on with missing values and coercion. The funny thing is that only the avg, stddev, and variance built-ins and their dim_... versions would exhibit the problem because all input to them had to be coerced to floats. When a missing value was in the input the coerced value would not reflect the missing value. Now NclMultiDValData.c takes care of this.
A related problem ocurred in the type functions because they'd replace
the missing values with the default missing  value when a coercion was
called without a new missing value. This caused problems for the modifications
to the MultiD object.

Anyhow all works now. I've looked where _NclCoerceData is used throughout
the code and do not think these changes will affect normal operation.


	-ethan


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5282 10fe2890-8a40-4696-8b1d-7ccbb19c398e