Fixed some bugs: 1) Hopefully the problems with running in mixed cmap mode on 24bit Suns is resolved. There were an X graphics context and some pixmaps being created using the default visual depth instead of the depth of the visual that ngi is actually using. So the bug actually would have shown up on any system where the depth of the visual used by the application did not match the default visual. Now the actual visual depth is used to create these entities. Hopefully I have found all the instances like this, but there could still be some lurkin somewhere.
2) Fixed a problem that caused parts of the application to hang because
an internal flag in the XWorkstation wrapper object was being incorrectly
set when the colormap editor was mapped or unmapped, when it should have
depended on the mapped state of the XWorkstation window only.

3) Coordinate variables belonging to regular ncl variables are now
passed along correctly to the plot objects, and the dataprofile grid
now understands the '&' syntax: the forms var&coord and file->var&coord
are both correctly parsed.

4) The varpage preserves knowledge of its hlupage linkso when
it is moved to another pane in the browser or when it is "hidden" and
subsequently redisplayed.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5310 10fe2890-8a40-4696-8b1d-7ccbb19c398e