Had a UMR in NclCCM.c
Joel Norris found a bug in NCL file variable subscripting. He was using
named subscripting with a combination of vector subscripts, a range and
a single index, definitely not something done frequently. Anyhow the
code got confused with this combination and caused a core dump on Sun
machines and bogus data on winterpark.

	-ethan


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5345 10fe2890-8a40-4696-8b1d-7ccbb19c398e