Implements a number of generic class resources and eliminates many specific class resources. Most of generic class resources are now defined in View.h and are thus available to any View obect. A few are defined in DataComm.h because they apply only to data containing objects.
Also a fix (in combination with GKS changes) for the oldest hlu bug:
hlu00065. This bug caused a core dump on some systems if DISPLAY is not
set correctly when running basic07.

Also fixed a problem that showed up as an error message when running
example cn17. It had the effect of making the info label not appear,
even though the default is for it to appear. (Now it does appear, which
may not desired for this example. The example may need to be changed.)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5367 10fe2890-8a40-4696-8b1d-7ccbb19c398e