Well I think I found what was going on with Brian E's core dumps. The complexity of his scripts made it extremely difficult to track down even with purify. Anyhow here's the abridged version of what was going on.
He'd create a contour plot with no cnInfoLabelOn == False.

Then some other function would set cnInfoLabelString but would NOT
set cnInfoLabelOn = True. So when ManageInfoLabel was called the
following lines prevented the new cnInfoLabelString value from
correctly being copied into the contour plot's data structure.

        if (! ilp->on && (init || ! oilp->on))
                return NhlNOERROR;


This would leave l->contourplot.info_string holding a pointer
to converter memory which eventually was free'd. Then something
else, in this case the map plot would allocate memory which
contained the location of l->contourplot.info_string in it.
This resulted in a Freeing Unallocated Memory error when
the contour plot was destroyed and exectued the following lines:

        if (cnp->info_string != NULL)
                NhlFree(cnp->info_string);

To fix this I moved the lines:

        if (! ilp->on && (init || ! oilp->on))
                return NhlNOERROR;

after the following lines:

        if (init || ! cnp->info_string ||
            cnp->info_string != ocnp->info_string) {
                int strsize = cnp->info_string == NULL ?
                        strlen(NhlcnDEF_INFO_LABEL) + 1 :
                                strlen(cnp->info_string) + 1;

                if ((lstring = NhlMalloc(strsize)) == NULL) {
                        e_text = "%s: dynamic memory allocation error";
                        NhlPError(NhlFATAL,NhlEUNKNOWN,e_text,entry_name);
                        return NhlFATAL;
                }
                if (cnp->info_string == NULL)
                        strcpy(lstring,NhlcnDEF_INFO_LABEL);
                else
                        strcpy(lstring,cnp->info_string);
                cnp->info_string = lstring;
                if (!init && ocnp->info_string != NULL)
                        NhlFree(ocnp->info_string);
        }


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5389 10fe2890-8a40-4696-8b1d-7ccbb19c398e