I am modifying the following EZMAP examples: mpex02, mpex04, mpex08, and mpexfi.
These changes reflect the fact that the default value of the parameter 'RE'
has changed (necessary because the resolution of output devices is greater
than it used to be).  This change affects the meaning of the parameter 'DD'
(the distance between dots drawn by MAPIT is 'DD'/'RE' fractional units).
The default value of 'DD' has also been changed, but user programs that set
the value of 'DD' themselves (as these examples did) will have to set its
value differently to get the desired dot spacing.  Without this change in
these examples, the dotted lines are drawn using dots very close together,
making the resulting ncgm's much larger than they used to be.

Some of the other examples in this directory also produce somewhat larger
ncgm's now, because the default value of the parameter 'MV' has also been
changed (a point closer to the last point than 'MV'/'RE' fractions units
is culled).  The effect of this on small-scale maps (those that show a large
fraction of the globe on the same map) is to now leave in some points that
would previously have been culled.  This is intentional.  Users who wish to
cull the same points they would have before the last update of EZMAP may
now need to make the value of 'MV' larger.

For reference, the changes in default values of EZMAP parameters were as
follows:

  NAME    OLD VALUE    NEW VALUE
  ----    ---------    ---------
  'DD'           12           96    (distance between dots drawn by MAPIT)
  'MV'            4            1    (minimum vector length for culling)
  'RE'         4096        32768    (resolution of plotter frame)

The programmer documents will be updated to reflect these changes.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5393 10fe2890-8a40-4696-8b1d-7ccbb19c398e