Fixed some bugs and added code to help the GUI and to make the HLU library a little more error-tolerant.
Create: added a CleanUp routine that is executed if an object create
fails. It removes a callback that otherwise can result in a core dump
when the app object is detroyed. It also destroys all children that have
been created up to the point of failure.

DataComm and DataMgr: tweaks to allow a DataMgr child to be destroyed
if the creation of the parent fails.

PlotManager: fixed some problems with annotations that manifested during
the implementation of plot style functions.

ScalarField and VectorField: added code to make it an error to supply
a data array with less than two elements along either dimension. (Hope this
this does not cause problems.) I am wondering if I should bump the number
up to 3, since the irregulartransformtion code requires 3 elements per dim.

Transform: corrected the _NhlTopLevelView routine.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5528 10fe2890-8a40-4696-8b1d-7ccbb19c398e