This is a set of modifications to better encapsulate the functionality of the shaper so that it can be used in a popup context. Some improvements have been made to its editing capabilities and it maintains its state better as well.
Also a small modification to the htmlview object: the text should not
be so big on Suns now. Hopefully it has not become too small for some
configurations. If anyone notices that it is too small on their
workstation (the help page text and the resource description text) please
let me know.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5552 10fe2890-8a40-4696-8b1d-7ccbb19c398e