Fixed a weird bug reported by M. Haley. The problem ocurred when both NCARG_ROOT and NCAR_LIB were set and _NGResolvePath was called from withing NCL i.e.:
load  "$NCARG_ROOT/lib/ncarg/nclex/gsun/gsn_code.ncl"


The bug resulted in several free memory reads. The problem is that once
_NGGetNCARGEnv finds an element in a specific directory heirarchy it doesn't
continue to fill out it's database of environment variables. So when
InitBaseDB is called for "sysresfile" _NGGetNCARGEnv recurses until it finds
"lib". However it doesn't continue and fill in a value for "root". Later
when _NGResolvePath is called on the above string it doesn't have a value
for root and calls _NGGetNCARGEnv to find it. There is a huge problem with
this approach. Namely _NGGetNCARGEnv calls _NGResolvePath. _NGResolvePath is
NOT reentrant because a) it uses static memory to build the path, which it
initializes on each entry and b) it uses strtok which  is absolutely not
reentrant unless compiled as such.

Due to the convoluted nature of the _NGResolvePath function I chose not to
mess with it and just fixed the problem of _NGGetNCARGEnv not filling out
the directory heirarchys. This fixed the problem but does not fix the problem
of _NGGetNCARGEnv calling _NGResolvePath. I am convinced the core dump
will not reoccur since all directory heirarchies will be followed no matter
what leaving no need for _NGGetNCARGEnv to ever recall _NGResolvePath.

-ethan


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5553 10fe2890-8a40-4696-8b1d-7ccbb19c398e