Added a font width multiplier to the xapp data. This allows for a better spacing for text in the Grid Widgets on Sun displays.
The shaper in the varpage is now updated (edits are completed) whenever
a new plot is created from the plot style menu, and whenever the browser
update button is pressed.

Added code to plotapp.c that recognizes functions and function parameters
in plot app resource values. It's not yet being used.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5649 10fe2890-8a40-4696-8b1d-7ccbb19c398e