Plot style functions are now parsed for their parameter values. The plot page grid listing the functions has been replaced with a new funcgrid that with an edit button in the column following the function name. Clicking the edit button pops up a dialog with a tree containing the function parameter name and an editable field containing the current parameter value. Right now all parameters are editable, even though some of them, such as the graphic objects should be made read-only. This will be addressed later.
A new plot style resource has been added. Actually a resource of a resource.
It is called ConformalGroup. It is an array resource that is applied to
the "*ndvData" resource. For example:

*ndvData@ConformalGroup : (/ $u$, $v$, $sf$ /)

This indicates that the three data items listed are required to have the
same rank, dimension names, and dimension sizes. Like "ndvUpdateFunc", their
can be multiple ConformalGroups, indicated by adding digits to the end of
the string. "ConformalGroup" without any digits following is a synonym for
"ConformalGroup0".

As usual, there are some miscellaneous bug fixes included in this check in.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5689 10fe2890-8a40-4696-8b1d-7ccbb19c398e