ContourPlot: added locally modified version of the CmpFAny routine (cnCmpFAny) that does not treat numbers close to 0.0 as the same as 0.0. This helps for picking contour levels in cases where the data consists of small values (order of 1e-8 or less). Hopefully it does not break anything else. Added type names for const field and no data error conditions.
Error: added a few selected error type names in order to allow certain error
types to be treated less severely in the GUI. These include the ContourPlot
constant field condition and the vector zero field condition and the no data
condition.

StreamlinePlot, VectorPlot: The zero field and no data errors are now named
type errors.

TickMark: This has the most extensive changes: added 8 new read-only resources
to help with autostriding based on label string size: tm..MaxLabelLenF and
tm..MinLabelSpacing. The first is the NDC len of the maximum len label for
an axis, the second is the min distance between label locations. Also
added another localized version of the CmpFAny routine. This one has a
new parameter specifying a minimum non-zero magnitude, allowing the
TickMark to adjust its concept of zero based on the axis extent. This
should fix the problem of TickMarks going away when the tickmark values
are very small (< 10e-7). This routine also uses a new method of comparing
numbers that is faster than the old way. I think that tests indicate it to
be equally accurate, but be on the lookout for any anomalies that
indicate otherwise. Another change is that if the TickMarks do go away,
because, for instance, the data extent is 0.0, it is no longer permanent.
You do not have to reset the tm..On resource to get them to go back on
when a new data field is provided with non-zero data extent.

XyPlot: fixed a problem that was preventing the data extents from being
recalculated properly when new data was supplied and segment drawing was on.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5778 10fe2890-8a40-4696-8b1d-7ccbb19c398e