Fixed some problem with editing data fields: if the user entered a bogus value in the plot page datavargrid cell, the GUI was marking resource functions that used the data as bogus forever. The user could not correct the situation. That is now fixed.
The globabl update button was not always working when data was changed (as
opposed to a View resource). That is hopefully fixed now.

Added processing in the ncl editor to recognize certain error types that
for which it is not necessary to notify the user (because the user is
notified by the plot graphic itself.) These include the no data, constant
field and zero field errors.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5779 10fe2890-8a40-4696-8b1d-7ccbb19c398e