PlotManager: add code to eliminate some unecessary calls to the TickMark setvalues method.
TickMark: a few cosmetic changes.

XyPlot: modified the setvalues routine to do the XyResetExtents call only
if a change to the extent has (most likely) happened. This eliminates a
circular situation where the extent is modified when resources not affecting
the extent are set by the user. (Well, something like that, it's a little
more complicated to describe in detail.)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5781 10fe2890-8a40-4696-8b1d-7ccbb19c398e