LabelBar: added code to keep the lbMaxLabelLenF and lbMinLabelSpacingF resources uptodate when a view-change-only setvalues occurs.
PlotManager: added new resources pmLabelBarKeepAspect and pmLegendKeepAspect. Only the
labelbar version is implemented so far. This allows you to specify that the shape of a
labelbar should stay the same when the shape of the base plot changes.

TickMark: fixed a uninitialized memory bug that purify pointed out.

Transform and TransObj: typed some error conditions with the NhlZEROSPAN error type name.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5797 10fe2890-8a40-4696-8b1d-7ccbb19c398e