graphic: added a function that returns the graphic array variable (the plot variable) given a scalar graphic variable that belongs to the plot.
xinteract: modified the interactive (rubber box) resizing code so that it now performs
an update on the plot. This allows plot style functions that might change the plot
appearance based on the plot size a chance to execute, possibly resulting in a new draw
of certain elements of the plot rather than simply a segment draw.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5810 10fe2890-8a40-4696-8b1d-7ccbb19c398e