This checkin replaces all instance of c_sflush in the HLU library with calls to c_plotif(0.,0.,2). This should flush the point buffers to the GKS layer, without doing an update workstation which unnecessarily causes metafile records to be output before they are full. Some extra calls to c_plotif(0.,0.,2) have been added for safety, but this call has much less overhead than sflush, and so should not impact performance.
Also modified some set calls to avoid unnecessary changes to the
viewport, which result in extra CLIP RECTANGLES being sent to the metafile.

The effect generally should be a decrease in metafile sizes. (But note
that other recent changes have in some case increased metafile sizes, so
comparisons are not necessarily straigtforward.)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5848 10fe2890-8a40-4696-8b1d-7ccbb19c398e