Primarily changes required by the zoom implementation in ngi; but also fixes a problem with the new float compare routine that manifested itself as bug #hlu00136.
After fixing the float compare functions that had been implemented statically in
ContourPlot and TickMark, I consolidated the two (identical) routines under a new name
"_NhlCmpFAny2" and moved it into the cmpf.c module where it belongs as a global function.
Modified ContourPlot,TickMark and VectorPlot to use this function.

Added 2 MapTransformation resources: mpDataMinLonF and mpDataMaxLonF. These are needed in
order to provide control over the cyclic longitudinal data range when NhlNDCToData is
invoked at the user level.

Modified the tfDoNDCOverlay resource so that it can be set as a 3-way enumeration, rather
than simply as a Boolean. (Boolean values still work as they always have however). The
first 2 enumerated options correspond to the existing Boolean options; the third option
is to accomodate NDC overlays on a MapPlot with the mpShapeMode set to FIXEDASPECTNOFITBB.
This seems to be the best mode to use for zoomable plots, because the viewport remains
constant and does not suffer from the diminishing size syndrome when the aspect ratio of
the plot constantly changes, as is typical for the zooming interaction using the rubber box.

Made extensive modifications to the PlotManager and Transform code to adapt the NDC overlay
concept to plots that occupy only a subregion of the viewport.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5930 10fe2890-8a40-4696-8b1d-7ccbb19c398e