Changes to support NCL operating as a DODS client. Note that actual operation as a DODS client requires substitution of the DODS NetCDF client libraries for the normal NetCDF library.
NclFile: file paths that start with "http://" are assumed to be DODS
data requests. These requests are handled as if they were normal NetCDF files.

NclNetCdf: a DODS-specific error message is generated if a path beginning
with "http://" cannot be opened. In order to improve performance for
data access over the net, all attribute and coordinate values are  now
cached when a NetCDF file is opened. Note that the caching is enabled (and may
have a minor positive performance impact) even for normal locally accessed
NetCDF files. Of course, there is some cost, hopefully minor in the usual
case, in terms of executable size.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5951 10fe2890-8a40-4696-8b1d-7ccbb19c398e