nclapi: this is a GUI interface to the nclapi intended to eliminate some redundant code, and clean things up a bit. It is only partially complete now.
The other files are changed to make use of this new interface.

mwin: a bug in managing the view object list was fixed. It was trigged
by an attempt to add a plot as a overlay to another plot that was an
overlay itself. (The HLU lib allows this, and simply passes the request
on to the base plot of the overlay, but the mwin view list management did
not handle this possibility.)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@5961 10fe2890-8a40-4696-8b1d-7ccbb19c398e