Fix for hlu00139. This was kind of interesting.
I think that this bug was introduced when Jeff added code to sort the
resource lists for improved performance (and maybe for another reason that
I forget now):

In a few cases, resources were being initialized using procedures
(NhlTProcedure initialization mode) whose results depended on the initial
values of some other resource. Originally, you could ensure that the
resource on which the procedure code depended was already initialized
simply by placing the dependent resource later in the resource list,
because resources were initialized in order of appearance. Even after
the sorting (by numerical order of the quarkified resource name strings)
was implemented, this usually worked, because the quark values are
assigned in ascending numerical order as strings are first encountered
and quarkified. In hlu00139 however, the "tiMainOn" string is encountered
and quarkified by ncl prior to the creation of a Title object (even
though the code where it appears is never executed!). This means it gets
a smaller quark value than the rest of the Title resources. Resource
sorting puts it ahead of tiMainString in the resource list and therefore
tiMainOn is initialized before tiMainString through a procedure that
depends on knowing the initial value of tiMainString.

The solution was to change the procedure to the standard one that simply
sets a flag (main_on_set) and put the dependency code into TitleInitialize
by which time tiMainString definitely has been initialized.

I found a few other cases of the same bug, which are also fixed in this
checkin.

"No procedural initialization that depends on the initial value of another
resource" needs to be a rule for coding HLU classes.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6072 10fe2890-8a40-4696-8b1d-7ccbb19c398e