Here's the rest of my last checkin's log:
Fixed a load error discovered by Denis. If a load was place within begin/end
then and calls to undef a common approach to avoiding multiple define errors,
would happen *after* the function or procedure was defined. Now I use the
instruction FPDEF to hold the function and procedure info and restore it if
it has been deleted. This has the affect of making sure the procedure isn't
recognized as defined untill execution gets to the point where it was
defined in the source code.

fixed minind and maxind in last checkin too.

-e


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6113 10fe2890-8a40-4696-8b1d-7ccbb19c398e