I have improved this routine to trim blanks from both ends of an incoming numeric label and to trim off any "1" or "1." or "1.0" or "1.00", etc., that precedes the "E" signalling an exponent (so that the resulting label will be of the form "10 [superscript] n", without the unnecessary "1 x" preceding it).
There's a "feature" on some machines (SGI's, for example) that will cause
this routine to fail to carry out its intended mission.  On such machines,
using the default format '(E10.3)' can create something like " -0.100E-2"
and this routine is not smart enough to realize that could be expressed as
"10 [superscript] -3".  On such machines, you just have to change the format
that is used to include a "1P" preceding the "E".

So now GRIDAL has taken a step in the direction of becoming AUTOGRAPH ...
which is how AUTOGRAPH came to be ... sigh ...


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6129 10fe2890-8a40-4696-8b1d-7ccbb19c398e