Fixes some bugs and implements sub-cell-size positioning for non-smoothed raster contouring.
ContourPlot:
Added resource cnMissingValPerimGridBoundOn, which allows a user to
specify that missing value boundaries that are coincident with the
grid boundary should be drawn.

Fixed #hlu00140 a problem with the initialization of an array associated
with the cnLineLabelStrings resource.

Modifications to allow raster contours to be positioned more accurately.
Also fixes to the code that determines the number of cells in a discrete
raster contour.

MapTransObj.c:
Some changes were required to allow raster cell counting and positioning
to be performed accurately when map overlays were involved.

Also implemented a routine that to a 3 or 4 digit degree of precision
returns the min and max lat/lons for arbitrary map regions in any
projection. The user can get at these values by performing a getvalues
of trXMin/MaxF and trYMin/MaxF.

Workspace.c:
modifications to _NhlCpcica to support the changes to the raster
contouring. Specifically, multiple calls to the GKS cell array function
are required in order to implement arbitrary clipping at the viewspace
boundary without actually drawing outside the boundary. This was a
requirement for accurate sub-cell-level positioning.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6136 10fe2890-8a40-4696-8b1d-7ccbb19c398e