Added Boolean resources wkSuppressBackground and wkSuppressBBInfo.
These resources together provide an interface for the new LLU parameter
'SU'. I have confirmed that the SU parameter gets set properly and that
the output PostScript is indeed different for each case. But it is not
obvious to me what the difference in the files is, so I don't know for
sure if everything is working the way it should be.
Let me know if you think the resource names should be different or if
it would be better to combine them into a single resource -- I could easily
change things at this stage.
 -dave


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6158 10fe2890-8a40-4696-8b1d-7ccbb19c398e