This fixes the core dump reported as bug #hlu00142. The core dump was caused by a incorrect printf specification for the error message -- a %s for which there was no string to substitute.
The fix does not allow the test code to run without error however. The
error is a result of the fact that coordinate variables are stored as
floats within the HLU library. There is not enough precision in the
float type to handle the double time coordinate variable used in the test
code. The IrregularTransformation expects the min and max values in its
coordinate arrays to be unequal within a precision of 5 digits. I am reluctant
to increase this to 6 digits (the most you could do with any hope of
reliability for a float) because I am afraid it would cause breakage
in some code somewhere.

The real solution to this problem would be to use doubles in all the
code concerned with coordinate variables throughout the HLU library,
but that is a rather substantial undertaking and the code would need
to be analyzed first to make sure it could still work with float
versions of libncarg and libgks (unless of course double versions of
those were used as well).
 -dave


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6170 10fe2890-8a40-4696-8b1d-7ccbb19c398e