Added functionality to make it easier to figure out the total bounding box of the drawable objects on a Workstation.
Added 2 read-only resources to Workstation: wkViews and wkTopLevelViews.
wkViews contains all the View objects that have been created on a
Workstation. wkTopLevelViews contains only the views that can be drawn
directly (i.e. views that are not overlays or annotations).

Also modified NhlGetBB so that it can alternatively accept a Workstation id
in place of the View id argument. In this case NhlGetBB returns the cumulative
bounding box of all the View children of the workstation.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6181 10fe2890-8a40-4696-8b1d-7ccbb19c398e