This modifies the PS workstation to allow the device bounds (DeviceLowerX, etc.) to be modified in a SetValues call. The change is not applied until just before the first workstation activate after the workstation was last cleared, so it cannot change in the middle of a frame.
I note that the test case I ran with two frames, in which the device
space was set to a smaller rectangle for the second frame, seemed to
confuse ghostscript. On my machine it created a smaller window than
normal for the plot, and did not show the complete picture for either
frame. Ghostview, however, worked fine.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6186 10fe2890-8a40-4696-8b1d-7ccbb19c398e